plugins/ks-test/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drift-ks-test"
version = "2.0.0"
description = "KS Test drift detection plugin — multiple testing correction, reference update, outlier handling, ECDF visualization"
requires-python = ">=3.11"
dependencies = [
    "drift-framework>=0.1,<1.0",
]

[project.entry-points."drift_framework.plugins"]
ks_test = "drift_ks_test:register"

[tool.setuptools.packages.find]
include = ["drift_ks_test*"]

[tool.setuptools.package-data]
drift_ks_test = [
    "web/templates/**/*",
    "web/static/**/*",
    "examples/*.csv",
]
